requirements.txt
fastapi[standard]
bcrypt
passlib
SQLAlchemy
psycopg2-binary
pydantic_settings